Moved camera quad calculations out of OpenGLDisplay.cs
